Fix dpl_openwrite function
The function had a easily reproductible bug that may prove a bit
unnerving. When trying to create a file in a subdirectory that already
existed, with a path containing slashes, the directory itself was
rewritten with the content of the file.

For instance, with the dplsh util :
:> mb testbucket
:> testbucket:
testbucket:> ls
testbucket:> mkdir dir
testbucket:> ls -lA
           0 2011-03-22 14:59 dir/
Total 0.00MB Price $0.000
testbucket:> put file.txt dir/file.txt
testbucket:> ls -lA
         256 2011-03-22 15:00 dir/
Total 0.00MB Price $0.000
testbucket:>

The bug was caused by a bad management of the directory search in the
path of the file. See the diff for more information.