droplet/dplsh: with S3 backend, doing "cd bucket_name" sticks the command-line
The main problem was in the wrong handling of blacklist address in case of virtual hosting mode.
With an invalid bucket, we ended invalidating the only address we used, thus resulting in a very
dire situation were no operation can be performed.

We simply chose to deny address blacklisting in vhost mode