Don't fail test when device not available.
There is a design flaw in the require_tape and require_changer functions
in scripts/functions it exits with 1 when you don't have a tape or
changer device defined. This is wrong and it should just like the
require_linux function exit with a 0 and just skip the test when it
doesn't have the right config. This way you can always run scripts like
nightly-all and don't worry about tests failing due to missing tape or
changer devices.