Remove unneeded estimate cmds from test.
When running tests with REGRESS_DEBUG the construct with redirecting the
output of a estimate listing gives to much spurious output. This code
probably originates from the time when the original script was designed
and it was copied to several tests afterwards. As we do an actual backup
in these tests the whole estimate doesn't add much to the test and we
have a seperate test which tests the working of the estimate command.