Tweak status scheduler output to cope with locales.
The %a and %b are locale specific strings which vary in lenght on each
locale. So we now keep track of the longest date string so we can align
the header right. For this to work we print everything into a temporary
buffer and keep track of the length and dump the whole output in one go
with the correct header spacing.