Send the catalog backup bootstrap file by mail
Having the bootstrap of the backupcatalog job is crucial
for disaster recovery.
Unfortunately, the default config didn't do that until now.

Now we automatically mail the bootstrap file of the backupcatalog job.

Also, we have now extended the BackupCatalog FileSet to include the
sysconfdir.

Fixes #24: Automatic sending of BSR (at least of the Catalog DB) via email