Add copies to the allowed keyword for restore.
Its seems that you can use a set of copy jobs to do a restore when you
specify the copies keyword on the restore commandline.

e.g. restore copies will select instead of the normal
full/differential/incremental jobs the copies of those Jobs. When doing
offsite SD-SD replication this can be a good selection criteria when you
lost your primay storage daemon but can restore from the secondary
or standby storage daemon. As long as you have a working database this
is a good DR procedure. You still need a working director and database
but as the database can be replicated as an HA solution and a director
also have a cold standby this could be part of a solution.

Fixes #36: Implementation of restore using a set of copies of an original job