Better handling of existing config files
When existing configuration files are found, in interactive mode the
user is asked if the existing config file should be kept and the newly
generated file is stored aside as config.new.

Also, the user can choose to move the existing file to config.old, and
install the new one in the original place.
This is what is done in silent installation mode.

The default for the uninstaller is now to keep existing config files.
In silent mode, the uninstaller will remove existing config files.

Fixes #140: Winbareos installer renames old config when updating

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>