mig_jcr is freed to early.
Because the mig_jcr is already freed in the migration_cleanup() function
all info is lost when we get to the after runscript. This is a design
error so we moved the freeing of the mig_jcr now to just before the Job
actually exits and has done all its work.

Fixes #139: export command in a RunScript ressource does not get the
            correct volume when doing scheduled JobType Copy