Implementation of an allowed job command keyword
Implement an allowed Job command keyword in the filed that sets which
of the different Job commands we should accept. Currently the filed
will execute any Job command unless people specify the -b or -r option
(backup only/restore only) which makes it a serious security concern by
much of the bigger customers security officers.

This new keyword implements a per director and a global list which is used
as a fallback when a specific one for a specific director is not configured.

We won't filter all the commands issued by the director only
the important ones. Things which we allow to be filtered are:

- backup
- restore
- verify
- estimate
- runscript

Fixes #32: Implementation of an allowed job command keyword