Cleanup of compression framework
As part of the integration of the LZO2 compression a new set of data
streams was introduced. Each compression stream now starts with a
so called compression header which says what compression was used.
We should deprecate the old GZIP datastreams e.g. still allow them to
be restored but not use them from now on and implement GZIP compression
using the same compression header which will cleanup the code enormously
and as we are planning on adding additional compression methods it would
be nice if all compression uses the same coding style and we deprecate
the old and non flexible GZIP datastreams.

Fixes #34: Cleanup of compression framework