Allow BAREOS to interface to multiple crypto libs.
OPENSSL gives all kind of headaches as its not compatible to AGPLv3 and
as such you always need a linkage exception in your LICENSE. Allowing
to interface to GNUTLS (LGPL) or Mozilla NSS (MPL/GPL/LGPL) makes this
probem much easier to swallow. This is the first work to allow either
one of the following crypto/TLS frameworks:

- None
- OpenSSL
- GNUTLS
- Mozilla NSS

Also only link and include the crypto specific headers and libs in
the place that actually defines the abstraction layer and not all over
the code.