Rewrite btraceback.
Rewrote btraceback to be a uniform script that behaves the same on each
platform. We now let configure determine which debugger is available on
the platform we are on. On execution of the btraceback script it checks
which debuggers were available and also checks if that one is available
on the current platform. If not it will fall back to runtime detection
of any available debugger.