Fix printing of crypto cache.
Also fixed input to always clean the input buffer so we don't read
garbage and added a -r option to be able to reset the expiry time of
all entries in the cache by setting the time they were added to the
current time. Updated manpage to document the new flags.