Add some safeguards for checking if we have a director.
When we never authenticated with an director the jcr->director pointer
will be NULL. When we then try to run the after runscripts (which are
not there also as they are passed in by the director we crash. So added
some extra safeguards and don't call the runscript at all when there
are not set.

Fix #103: bareos-fd 13.1 on ubuntu 10.04.4 crashes with segmentation violation