Rename global resource in director,
Rename the global director variable to the same name used in the storage
daemon and filedaemon which is me. Also drop the two global variables
for the FD and SD timeout as those are part of the global resource
anyway so we can reference the me variable and don't need to polute the
global naming space of the binary with them.