For migration/copy Jobs need to look at mig_jcr.
Added new %V option to expansion variables which is the same as %v
but always prints the destination Volumes a Job used. For migration
and copy Jobs this expansion looks at the mig_jcr pointer in the JCR.

As this mig_jcr only exists in the director the %V is expanded in
the callback expansion function in the director as in the generic
expansion function is compiled as part of the generic shared library
we are missing that variable there.

While implementing this expansion also reformated the whole function
as the switch statement was wrongly indented. Removed unneeded break
statements after a return. Document what is expanded in the callback
expansion function.

Fixes #139: export command in a RunScript ressource does not get the
            correct volume when doing scheduled JobType Copy