Extend cancel command to be able to cancel all running jobs
Especially when testing, it would be a very nice feature to be able to
cancel multiple jobs at once.

For the first step, it would be nice to be able to cancel all jobs.
Also useful would be to be able to cancel, jobs of a certain state, e.g.

cancel state=waiting

Fixes #144: Extend cancel command to be able to cancel all running jobs