Added scheduler preview
The new command "status scheduler" shows the status
of the bareos scheduler.
The output first shows what schedule triggers what
jobs, then the scheduler executions of the next
14 days are displayed including the overrides.

By giving the param "schedule=" the display can
be limited to a certain schedule and by giving
"days=" the timeframe for the preview can be
set.

Also, the dotcmd schedule was added, so that the
tab-completion works for giving schedules.

Fixes #14: Implement a scheduler preview for up to one year.