Drop syslog.target from systemd service definitions.
Newer systemd versions implement a internal binary journalling system
which replaces syslog so the syslog.target is no longer available in
newer versions of systemd. So we dropped the dependency on it and hope
that our other dependencies make sure the syslogging is available.

Fixes #203: systemctl start bareos-fd.service fails on Fedora 19