Fix backup summary.
Always lookup the media record of the last used volume after calling
db_get_job_volume_names() so we always print the right info and not
lookup the media record based on the current content of jcr->VolumeName
which works only for normal backups but not for virtual fulls as there
it will have the name of the last volume that was read for creating the
virtual full. Looking at the migration code it also does a lookup
of the last volume name returned by db_get_job_volume_names() there.

Fixes #211: Wrong data/output/calculation in log of "Last Volume Bytes"
            and "Volume Session Time" when doing Virtual Full Backups