Fix restore handling of NDMP when using wrong storage.
For a NDMP restore its needed that the storage passed into the restore
job is a storage definition with paired NDMP storage. When this is not
the case the pstore pointer will remain NULL and we crash in the
build_ndmp_job() function. Now we check if we have paired storage
and also changed the cleanup somewhat so we don't try to close the
bootstrap file etc when we didn't open it (which also lead to a
segmentation fault.)