Some keywords are only allowed in fileselection mode.
We now always send the BNET_START_RTREE and BNET_END_RTREE signals
even when the api mode is not enabled. This way bconsole also knows
when its in file selection mode and the tab expansion can show the
only expand keywords allowed in a certain mode. e.g. no more cd
and ls expansion when not in file selection mode which leads to
the dreaded "*cd .lsdir: is an invalid command." and
"*ls .ls: is an invalid command."