Sortable tables in bRestore
Quick as is fix. Enable sorting in bRestore by setting the sorting
attribute using setSortingEnabled() on the bRestoreTable objects.
For FileList and FileRevisions we disable the sorting during the
filling of the table (to not make it to slow) and explicitly sort
the FileList on the filenames in them ascending and the FileRevisions
on the revision date descending (so you get the newest revision first
in the list.) Some sorting doesn't work or doesn't work right as we
need to write new sorting operators which are not there. That is
something we won't solve any time soon. The restore list also has
the sorting attribute enabled so you can use sorting in that view too.

Fixes #215: Sortable tables in bRestore