Fixed rights on tray-monitor.conf
When limiting the access to the config files to admin only
we also did this to the tray-monitor configuration file.

Of course this is wrong, as the traymon is started as normal user.

This patch fixes this and sets the tray-monitor.conf file ownership and
access rights to the group "users".

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>