Allow upgrade on windows
Now we check if bareos is already installed on the windows machine.
If so, we call the uninstaller of the previous version and install
ourselves afterwards.

The newer uninstaller keep the configuration files as the uninstaller is
called accordingly.

The dialogs for parametrizing the configuration and the dialog
for the config snipped are not shown, as the existing configuration
files will be used.

Still the user has to choose the same components during the upgrade
as were installed before.

fixes #225: Windows installer can't upgrade in place