Fix problem with copy/migrate jobs.
Now we select the right catalog and not always set a pool override of
the catalog we run into a crash in set_jcr_defaults as that now
derefences the job->client pointer which for copy/migrate jobs may not
have been set. This patch adds an extra catalog override method to
allow you to set a job specfic catalog setting. If none is set the
default (first) catalog is choosen for the copy/migration Job.