Fix bug in dpl_condition_t structure
The dpl_condition_one_t data structure has a field which is meant to
hold the value of an etag, i.e. the value returned from an HTTP server
in the Etag: response header and passed back in an If-Match: request
header.  This is typically an MD5 hash of the file in ASCII format (but
not necessarily, RFC2616 is vague about it's actual format).

The field was declared char[MD5_DIGEST_LENGTH].  Which would be great if
the field were actually handled as a binary value - MD5_DIGEST_LENGTH is
16, the length of a binary formatted MD5 checksum.  Unfortunately the
code treats it as ASCII formatted, so the correct length to avoid
truncating the value in dpl_parse_condition() is 33.

Elsewhere the dpl_sysmd_t data structure also stores an etag, as a
char[64] which is an uninterpreted string.  Given the RFC's vagueness on
the format, this makes more sense to me, so I've changed the
dpl_condition_one_t data structure to do the same.

Note: this is an ABI change.  I'm assuming we can get away with that.