tape examples now use bigger block and file sizes
As the defaults for Block Size and File Size are much too small
for modern tape drives, we now set some adequate defaults in the
example configs.

Also we link to the tape speed whitepaper where the effect of big
block sizes can be seen in detail.

Fixes #248: default values for block and file size for tape devices in
configurations are too small