Allow history file and size to be configurable.
Up until now the readline history file always had a fixed name and size
of 100 entries in memory and 200 on disk. You could always configure a
name for the history file but it was never used. We now also always
truncate the history size to the configured size before reading it into
memory.