Make the error message a bit more explicit.
When the compat mode is enabled we report a Bacula compatible version id
for the filed protocol. But if you want things like passive mode or
plugin passing you need to disable the compat mode. Changed the error
message issued when using passive mode or plugin option passing when
talking to a Bareos filed in compat mode. This way the use a least knows
that it could also be a problem with the compat mode.