Upgrade bpipe plugin to use new plugin config parser.
Now that we use a new config parser for the mssqlvdi and python plugin
it makes sense to also let the bpipe plugin use this code. For backwards
compatability it also still supports the old plugin definitions so you
can also still restore older backups with the new plugin. Extra support
is added so that plugin options passing now also works and allows you to
override the plugin parameters.