Add bEventNewPluginOptions event.
When pushing explicit plugin arguments use a new event
bEventNewPluginOptions instead of reusing the bEventPluginCommand
event which is also used when parsing the fileset when doing a backup.

Most plugins should now also allow multiple plugin lines within one
include block of the fileset which was broken when it used the new
config parser. In combination with forcing certain explicit restore
options by the director using the plugin options passing this may lead
to some surprises however so when using specific plugin options its wise
to only restore one file at a time to be sure the right restore options
are used.