BVFS reported a problem for <id> fix.
There was a mixup of the return value returned from the
bvfs_update_path_hierarchy_cache() function which returns a zero which
is interpretted as a false by the calling dot_bvfs_update() function.
Lets return a proper boolean value and only set that return value to
true when we leave the bvfs_update_path_hierarchy_cache() function due
to hitting either an empty jobid list or after processing the jobid list
successfully. Any other exit (due to joblist corruption or failing to
update the path hierarchy cache) leads to the function returning false.

Fixes #217: BVFS reported a problem for <id>