Copy job definition requires unnecessary settings
Due to implementing this feature we can no longer asume that there
always is a Client and Fileset in a Job definition. So we need to learn
the .defaults cmd that is needs to print *None* when we don't have a
Client or Fileset and not just dereference a NULL pointer.

Fixes #220: Copy job definition requires unnecessary settings