Rework database scripting.
Having the actual DDL (Data Definition Language) in scripts is not very
flexible so we move all DDL into seperate files. We also dropped all
database specific scripting and as such the generic script now knows
for all supported database types what to do on create/update/grant and
drop.

You can also have a read-only user on your database that is allowed
to select stuff but not anything else that might be interesting for
Management Queries.

The update script now can also upgrade from a database that is several
versions old. For now that means you can upgrade from a Bacula database
to the first Bareos database schema and also to the new Database schema
we are introducing with this patch. For the future it means the script
will perform all needed updates for you if its has the wanted so called
conversion available (e.g. the way to go from one version to an other.)

We also added more functionality to the bareos-config script and splitted
the function into a bareos-config-lib which can be included by scripts
so they can get config information like the working dir etc. This is
already used in all the new database scripts.

We also no longer use explicit absolute pathnames for database utilities
but make sure the utility can be found via the PATH, if your PATH already
provides the utility it will be used and otherwise the value will be used
that was determined at compile time.

Fixes #236: Hardcoded path for postgres (verison) in scripts