Implement generate_plugin_event bEventPluginCommand
Up until now the so called Options Plugin never did a
generate_plugin_event() for the bEventPluginCommand event when a Options
plugin was specified in the fileset. Now when we finished parsing the
fileset in the term_fileset() function we generate those events for each
plugin which is configured as a Options Plugin in the fileset.

Also added an extra check in the python plugin to check if the python
specific module is loaded before trying to access the dictionary in the
PluginIO() function so in the future we don't crash there anymore.

Fixes #281: fd crashes when trying to execute a python plugin as option plugin