Optimize checking of required and default values.
When printing the configuration we now first check if we need to print
the value by checking if its a required value or matches the default
setting in the config table. By moving it before the big switch we can
eliminate quite some duplicate code and always use the right type of
value in the union with result types.