Refactor status printing.
Try to use a POOL_MEM whenever possible so we also don't make mistakes
in the length of the status string a possibly overrun an string. Also
if we don't print a string because of being in API mode it doesn't make
sense to even build the status string (as we are not going to print it
anyway.) Also fixes 2 other Coverity high impatct bugs.