Storage to storage copy jobs don't work with TLS.
When both SDs think they are a TLS server the handshake will never be a
success. Now changed the code to use the initiate variable we have which
we already use to know who starts the challenge protocol. That way we
get one TLS server and one TLS client which should work much better.

Fixes #290: Storage to storage copy jobs don't work with TLS enabled.