Use dbconfig for Debian based distributions
dbconfig-common is a standard mechanism for Debian based distributions,
to assists applications that require a database.

The user still has the choice to configure the database manually, if not,
dbconfig detects what database backends are installed, and the user gets
asked, what connection parameter he wants to use.

If dbconfig is configured correctly, database schema updated are done automatically
on package updates. With export DEBIAN_FRONTEND=noninteractive the user don't get ask,
and dbconfig configures the database backend as good as he can.

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>