Sync md5 and sha1 code to what others use.
We have stuff that could make some distributions scream that things might be
not fully opensource so we sync here to what other projects use e.g. a
fully public domain version of sha1 and md5. When things are compiled
using OpenSSL we even fall back to their implementation of md5 and don't
compile an own version at all. The rest of the code was changed to use
the new context names and the small change in the function names.