Multiple "Exclude Dir Containing" Flags not honored
It seems when implementing the Exclude Dir Containing option in the
fileset unlike every other item it doesn't allow you to specify multiple
values for this config item. Which is kind of strange and would make it
somewhat better usable. This patch adds the missing support by just
storing the values in an alist just like all others. This also fixes a
memory leak which happens when you used to specify multiple entries as
then it would overwrite the previous string dupped value with the new
value without freeing the old value.

Fixes #311: Multiple "Exclude Dir Containing" Flags not honored