Fix 64Bit install dir problems
On 64Bits, there were problems when a different install location was
chosen.

Now we only set PROGRAMFILES to PROGRAMFILES64, if the INSTDIR was not
altered.

On uninstall, we do not set the INSTDIR anymore as it is set correctly by
the uninstaller.

Fixes 324: Windows installer (64 bits) overwrites user install dir
settings.