First prototype dynamic loading SD backends.
We now store all shared objects used in all SD programs
(stored, btape, bls ...) in a shared library named
libbareossd-<version>.so so the final binaries get much
smaller and share the same code. Next to that we started
the first work on making tape specific code abstracted in
the tape_device class and no longer in the generic device
code. Now we implement the different methods of the different
device classes using virtual methods some are pure virtual
which means that the derived class needs to implement this
method itself and some have a default implementation which the
class can override.

This also contains the first try at making it possible to
dynamically load the different SD backends. The file backend
will always be part of the libbareossd.so but the tape and
the different cloud storage backends are dynamically loadable
using the same technique as what is used for loading the
different catalog backends. This means that we can distribute
the different storage programs without the need to have all
kind of dependencies on the different cloud backend we support
now.

As part of this big refactoring we removed the vtape code which
doesn't serve a real purpose anymore in Bareos as there are much
better alternatives (MHVTL and QuadstorVTL) nowadays and we should
carry on unmaintained code. The alternatives also see steady
development while vtape hasn't been updates in ages. Also these
alternative VTL emulators implement a true autochanger, encryption
and much more like personality support for all major brands in
real tape hardware which also means we test our tape code much
more thoroughly.