Use a special backend dir to store both catalog and storage backends.
We want to store the catalog and storage backends in a special dir so
they are not in any lib directory as they are not really shared
libraries but shared object with some very specific entry points.

The code is now so that by default the configured backend dir set at
compile time is set as default and the user can specify additional
directories that should be searched for backend in the configfile of
both the director and the storage daemon. The code checks if the wanted
backend is located in one of the specified backend directories and it
loaded from the first directory it is found in.

The packaging is changed to do the following:
- Enable building packages with dynamic-storage-backends.
- Different backends are handled in separate libraries.
- The bareos-storage-tape package does now also contain the required libraries.
- The fifo backend is also handled by a separate package (bareos-storage-fifo).