PostgreSQL and MySQL compatibility.
Table- and column-names are now passed into the SelectObject by a helper function, which
converts the string to lower-case e.g. for PostgreSQL or leaves it camel-case for MySQL,
so we don't need to repeat most of the code for different databases, as identifiers might
be different depending on which database you use. This is a workaround for an inconsistence
in the Bareos installation scripts relating the creation of needed database tables for
the catalogue. The helper function strdbcompat() might also be extended in future for Oracle
databases, where identifiers are handled upper-case by default and possible other databases.