Change the socket constructors.
As all BSOCK_XX classes are a derived class of BSOCK it makes more sense
to initialize the shared members in the BSOCK constructor. As on
creation of a devived class the base class constructor gets called
anyway this will work nicely. We also don't have to intialize everything
to zero as the BSOCK class is a derived class from the SMARTALLOC class
and as such the constructor of that class will initialize the memory to
zero.