Enhance read device selection to respect media_type
When looking for a device to mount a volume that is to be read bareos
goes for the device the volume was last mounted in. If the original
device is blocked but there are other devices that can be used those
will not be selected and the job will fail.

Fixes #326: Enhance read device selection to respect media_type

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>