Proper fix for using fallback MD5 support.
If we don't use the openssl MD5 support on windows we need to fallback
to using a C++ version so we don't have a name clash with the normal md5
routines available. So if we use our public domain version of MD5 we
always compile it as C++ code from now on.