Give access to job size to the plugins + enhancements/corrections
The JobBytes, ReadBytes and LastRate values were not available to the
plugins of the Director.

Also, since the plugin needs to know about Canceled jobs, I moved the
is_job_canceled() function after it passed through the plugins.

While testing my changes, I came unto a signed/unsigned problem with
values >2^31 (0000002:0000002,100,000,000). Python does not sign its
integers so when passing an unsigned value to Python, it needs to be
made explicitly. I added a function to correctly send the added
variables (JobBytes, ReadBytes and LastRate) to Python as unsigned but I
also moved some variables sent as signed integer to Python while they
may be over 2^31 to that function (*JobFiles and *Errors) since they are
defined as unsigned in the C code.

Fixes #334: Give access to job size to the plugins +
            enhancements/corrections

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>