Ditch useless unions in dird config.
In essence they save a maximum of one pointer value which is not worth
the side effect it has for printing a config data. To be able to
determine what data is actual in the union you need an extra variable
describing what is the pointer about so then there is little value left
in using an union.