Fix runscript printing.
Keep track if a runscript is entered as a short runscript using the
short_form boolean so when printing it we also know if we need to print
it either in its short form or in the new long form. Also keep track
using the from_jobdef boolean if a runscript is inherited from a JobDef
as then we also don't have to print it in the dumping of the current
config.