Refactor socket handling.
Try to align the socket server in each daemon to work in the same way.
Extracted the code for this socket server into a seperate file which
dispatches to the right connection handler.

Added first steps for allowing different type of connections on the socket
the director has open. Analog to how the filed and stored determines based
on the Hello message what type of connections is established. For now
things should always connect as a normal console connection but the
logic for determining if a Client connects are implemented only there
is nothing done and the connection gets aborted right away.

Use same kind of refactoring that already happened to FD and SD for
allowing to reuse the two way challenge authentication for multiple
connections without recoding the same over and over again.