add configure option dynamic-debian-package-list
When building packages for Debian,
the resulting packages are defined in the file debian/control.
This file must exist at the beginning of the build process.
As Bareos is build for different plattforms, the resulting packages differ.
Therefore configure can overwrite debian/control
so that is matches the configured options.
However, the Debian project does not allow this file to be changed.
Therefore we make this feature optional.
The configure default is to keep the static debian/control file.
When configure is used with the option
--enable-dynamic-debian-package-list
debian/control will be overwritten with the template debian/control.in.