Automatic detection of volume mount points
Volume Mount Points are now automatically detected.

If we detect a volume mount point, we check wich volume is
mounted there and add it to the snapshotset.

If the volume is not already in the snapshotset, we add it and
check if the new volume also has volume mount points.

The total number of volume mount points and the number of volume mount
point volumes added to the snapshotset is reported, if VMPs exist.

Signed-off-by: Philipp Storz <philipp.storz@bareos.com>
Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>