Fix btape bug with big blocksizes
there is an overflow when calculating the maximum file size, which leads
to a eof being written to the tape where btape does not expect it.

Then during the read of the data the unexpected eof lets the test fail.

This fix corrects this overflow.

We also set a higher debug level on the block writings in block.c

Fixes #381: btape tests fail on blocksizes > 256k