ACL enhancements
- Add support for console profiles

This adds some basic support for defining a profile with some basic ACLS
which can be referenced by any console resource. This way you don't have
to define the same ACLs over and over again for each named console.

The way this works is that the ACL definition in the console resource is
checked first to see if it already allows you to access the wanted
resource. If that is not the case it will check if your named console is
connected to a certain profile and use the ACL setting in that profile
to get access to the resource.

There is one special case and that is the where ACL if you want to use
the where ACL in the profile you should set the where ACL of the named
console to something like *none* as an empty where ACL means allow
restore anywhere.

- Add support for regular expressions in ACLs.

This makes ACLs a lot more universal and usable.

e.g. you can now use a poolAcl which states: .*NDMP.* and get access to
all pools which match that regular expression (Thing are compared case
insensitive).

- Add support for deny ACLs

e.g. a command acl like this will work now: !show, *all*

Which will allow all commands but not show.