Add configure console command.
For now this is a dummy command but eventually this will implement
the interactive configuration of the director. For now we only use
this command to determine if we should print the sensitive data
when dumping resources in the show command. If you have a console
that allows either the configure or any command you will see the
sensitive data (passwords) otherwise not.