Ignore tape labels in NDMP code.
When we encounter some sort of tape label in the backup stream
(e.g. FileIndex < 0) then we should just skip that data and not
use the default handling which leads to an EOF. When we now encounter a
unknown data stream we also print an error including the number of the
unwanted datastream.