Use either the CreationTime or ChangeTime of the FILE_BASIC_INFO.
We switched to using the GetFileInformationByHandleEx() API call
when available to also get changes to directories when ACLs get
updated. It seems we need to take the highest value of the
CreationTime and ChangeTime value of the FILE_BASIC_INFO structure
to get a proper st_ctime stat structure field.