Fix removing of supplemental groups.
Need some coding as usermod -a doesn't work on all platforms.
We now figure out if we are creating a user from scratch e.g.
its doesn't exist at all or need to update it. For that we look
at all groups the user currently is in and add the wanted groups
if they are missing. This should work on all platforms as we
script around and only use options available on all platforms.
When we use usermod -G we supply the full list groups including
any specific groups added by the sysadmin.

Fixes #382: Don't remove supplemental groups on install.