Add special device options config field.
This adds a special field to the device resource where you can specify
specific device options. This is mostly interesting for the cloud
storage backends so you can specify the cloud specific parameters
including things like usernames and passwords. We don't want to store that
in the archive devicename as that is shown on storage status.