Explitly check password encoding for NDMP resources.
If you define for a client or storage daemon the protocol after you
configured the password it will store the password using the wrong
encoding as at the time of parsing it still thinks you will be using the
native protocol.

Fixes #408: Director crashes on password encoding for NDMP storage.