Refactor pretty printer of schedules.
Seems the pretty printer had some problems with printing some schedules,
Given that this is almost a programming language this is not to
surprising. Changed the code to first determine if all bits are set so
we don't do hard work of finding intervals if we are not going to print
the interval at all.

Fixes #407: print_config_run() doesn't print months
Fixes #409: Console Shows Incorrect Schedule

Conflicts:
	src/dird/dird_conf.c