Fix bug #184
In part of the codebase, calling functions to add some data consumes
the pointer given to it, and the developer must then avoid freeing the
resource. In this specific case, the memory isn't consumed, but duplicated,
and the resource given by the caller must be freed later on.

This was the matter at the core of a leak where the memory wasn't consumed
nor released.