improved JSON support
In order to better interact with other programs,
the Bareos Director console now supports
the ".api json" command, to switch to JSON output.
Currently only a subset of the commands are supported
(the ones required by bareos-webui).

To archieve this, Bareos can be linked against the jansson library.
Other parts of Bareos (output of the configuration options)
now also use the jansson library.