Added cmocka unit tests
First implementation of unit testing in bareos
using the cmocka test framework.

Therefore we built the cmocka libraries for
all platforms we support (also windows).

On Unix/Linux, the unit tests are executed now
during package build by calling
   "make check"
,on windows we package the test exes so
that they can be executed on the target
system during ci tests.

For the first we moved the existing
TEST_PROGRAM main() unit tests
of src/lib and src/findlib to a special
test that is being compiled in the subdirectory
unittests/ of each directory.

The specfiles were also refactored.

It should be easy to add unittests to other
directories in the same manner.