Don't crash when ua->UA_sock == NULL in prtit()
Some admin jobs have an UA context but not a ua->UA_sock and when you
then blind use ua->UA_sock->fsend() you crash in the worst possible way.
We now call ua->send_msg() which has some fallback logic when
ua->UA_sock is NULL and then uses Jmsg with M_INFO to redirect the info
to the Job.

Its might not fully fix #418 as its seems the admin Job want to have
interaction with the user which ain't going to work in an admin Job but
crashing is about the worse what can happen.

Fixes #418: traceback runnning after job