Fix problems on Fedora 19 and 21.
Glusterfs on Fedora 19 seems to be old and as such does not allow us to
build the gfapi-fd plugin. As Fedora 19 is not supported anyway it makes
little sense to add a lot of work to make the gluster stuff work on that
platform.

On Fedora 21 it seems the compile flags are more strict which leads to
the following error:

tray-monitor.cpp:65:32: error: format not a string literal and no format
arguments [-Werror=format-security]