Add a empty_folder_emulation flag in the profile
This flag is a flavor that was previously the only way to manage directories in
a S3 bucket. Now, two flavors exists, with the default one completely
simulating directories using the common prefixes in the bucket, and the second
one (which must be activated explicitly) matches the old behavior.

Note: This leads to an inconsistent behavior between Emulated and Non-Emulated:
 - In Emulated mode, opening a directory requires to remove the ending
   delimiter from the path (the backend accesses the empty object representing
   the directory, which name does not contain an ending delimiter)
 - In Non-Emulated mode, opening a directory requires the ending delimiter (the
   backend uses the ending delimiter to identify if it is a directory)

Adresses Issue #191