Fix some problems in unsupported database pooling code.
Although the pooling code is unsupported it currently crashes if you run
reload twice. We should not use foreach_dlist() in a loop when we remove
and free members of that list in the loop. Now we use first() and lookup
the next element using get_next() and store things in a temp variable
which we use in the loop to switch to the next element.