For a command plugin always use the no_read field of savepkt
A command plugin can set the no_read flag in the savepkt returned by the
startBackupFile() method in the plugin events. Up until now setting this
no_read flag to true would not make a lot of difference as the code only
checks for 'ff_pkt->cmd_plugin && !ff_pkt->no_read' to enable the
do_read variable but if the do_read was already set to true and the
plugin has set no_read to true then do_read will never be set to false
again.

These changes make the no_read flag from the plugin leading and always
use that setting (which is false by default) as the setting to determine
the do_read flag. Only for a non cmd-plugin we will determine the actual
reading of the file or not by looking at the specific file type.