Fix config pretty printer.
When printing the compression used it eats an extra char so the pretty
printer doesn't print the next option in the option string. We should
also print GZIP without \n as we print the level after it, currently
its printed on the next line.

Fixes #466: option noatime = yes in bconcole displays like 6 (digit)