Do not assume a filetype in digest_file
Unfortunately, the ff_pkt->type is not set correctly when
digest_file is called during accurate comparison.

We get the value that stored before in ff_pkt->type by the last
file/directory that was processed, not by what we actually see.

That is why we do always calculate the digest now.

Fixes #462: windows file daemon and zero-length files in Inc backup Description