Don't call PyCheckFile()
When we have a inactive python-fd instance in a Job e.g. no option or
file plugin definition in the fileset we are still called as part of the
Checkfile method which calls each plugin loaded. This gives problems as
we try to lookup the entry point into a non-existing python script and
as such we have no dictonary at all which means the lookup will
dereference a null pointer.

Fixes #473: Incremental backup failure