Revert "fix possible memory leak of resource_ue"
This reverts commit 8b2288c314bdd2b9debe38b2b0b7c5d57e58c872.

The patch was incomplete, as it only addressed the potential memory leak, and
didn't make sure that every "len" used by the whole library for
DPL_APPEND_STR() macro calls were actually size_t variables.

Hence a warning of using a lot of different pointer types implicitly casted
into size_t pointers. This is a risk that we do not want to take, and this
patch is reverted until a more complete fix is proposed.