Status storage should only show configured devices.
When you do a status storage from the director to a defined storage on a
storage daemon you expect that it only shows you the actually configured
devices in the director and not all devices configured in the storage
daemon. For that to work we now send the list of devices (normally one)
that are configured in the director config and use that as the selection
criteria for showing the right devices only.