Cleanup director and use bool instead of int
We have a lot of functions that return an int 0 or 1 which can
just as well return a proper boolean as return value. This patch
tries to cleanup this in more places as already done before so we
mostly use bool where possible and int only when things cannot be
solved by a bool e.g. when false and true are not enough.