Integrate messages to OUTPUT_FORMATTER
This patch integrates message handling into the OUTPUT_FORMATTER so
that messages (info, warnings or errors messages) can not degrade
JSON output anymore.

In JSON API mode, messages are now kept separate from the actual output.
Normally messages will not be send, except if one of the messages is
an error message.
In this case, the whole JSON result will be marked as error and the
current result plus the messages will be delivered as an error message.

Also this patch sets batch mode in api mode JSON.

In batch mode, commands do not show additional prompts.
If not all parameter are given, a command fails to execute.
This is the wanted behaviour for the API mode JSON.