Clear encryption and compression flags if unsupported.
Analog to how we clear the ACL and XATTR flags in a fileset if we run a
backup Job on a filed which doesn't support the wanted feature we now
also clear any compression or encryption flag. By clearing the flag
before we start the actual Job we no longer get those nasty errors
in which by the time we call blast_data_to_storage_daemon() we abort
the Job due to missing features.

Clearing compression flags should be no problem and a warning is logged
anyway. For the encryption not being available one could use the new
force encrypt flag in the fileset to fail the Job if encryption is an
absolute requirement.

Fixes #496: File daemon without compression support / director with
            compression enabled - weird error message