Add the guid cache pointer to the useragent.
From now on useragent has a guid cache to store the uid and gid to
name mappings.  The cache is freed when the useragent is destroyed.

Until now only the interactive ua_tree.c used a guid cache from
uid/gid to user/group mapping but we also want the BVFS api mode 2
(JSON) to have this information for file selection.