.default job: send each line as seperate message
The job tab in BAT calls "defaults job" in API mode 1.
It parses the result message by message, instead of looking for a
seperator.
This patch adds the OUTPUT_FORMATTER:send_buffer() function.
When called, it sends the current buffer as a seperate message.
Currently it is only used for the ".defaults job" command.

Fixes #505: BAT hangs on Jobs tab