Make the create_file messages of type M_INFO
If a file was not created during restore because of
the overwrite settings, this was never visible in the
restore job log because the messages were of type
M_SKIPPED, which is filtered by default because it
creates an enormous amount of output during normal
backup operation.

Now we changed the messag tpye M_INFO which means that
this info will be visible in the restore log.

If you set the overwrite to always you won't be bothered
by these info messages and we pretend you know what you
are doing.