Add support for restore to local VMDK.
For this we introduce a few new flags e.g. -l which means use
local VMDK e.g. don't try connecting to a Vsphere server and
a -C flag which creates a new local VMDK using the VixDiskLib_Create()
call. With the -d option you can pass the name of the VMDK to write to
without the need to change the JSON workfile you provide. Without the
-d option the VMDK name will be retrieved from the JSON work file.