Fix pattern matching crash on windows.
When we use the fallback bregex code we cannot use a nmatch == 0 and
pmatch == NULL as we can with regular regexec on UNIX nowadays.

One day we might want to teach this fall back code how things work in
reality in the real world now but for now we work around it.

From the current manpage on Solaris on regexec:

If nmatch is zero or REG_NOSUB was set in the cflags argument
to regcomp(), then regexec() will ignore the pmatch argument.
Otherwise, the pmatch argument must point to an array with at
least nmatch elements, and regexec() will fill in the elements
of that array with offsets of the substrings of string that
correspond to the parenthesised subexpressions of pattern.