Add support for VixDiskLib_PrepareForAccess.
We should tell Vsphere we are opening disks e.g. call
VixDiskLib_PrepareForAccess() before opening and VixDiskLib_EndAccess()
after closing.

From the include file:

This function is used to notify the host of the virtual machine that the
disks of the virtual machine will be opened. The host disables
operations on the virtual machine that may be adversely affected
if they are performed while the disks are open by a third party
application.