cleanup reload_config command
reload config command is callable as signal handler for SIGHUP
and by the bconsole reload command.

Make sure, that the SIGHUP handler is only called once.

Do not terminate the director in case of config errors.

Fixes #491: bareos-dir dies on reload if config has typos
Fixes #503: Config reload leads to director crash if a duplicate resource named is present