bcommand: fixes a problem in the output format
Bareos has different methods to generate output for consoles:
  * print
  * messages
    * send_msg, error_msg, warning_msg, info_msg

Strings from the print method are first stored into a text buffer and only send
to the console on request or at the end of the bcommand.
Messages are send directly to the console.

Some methods use both types to generate output.
To show them in the correct order, this change first flushes the text buffer
to the console before sending a message.