Add "accurate" support to VMWARE plugin.
As we want all files to be restored in the same order that the backups
are taken of the VM we have to fool the accurate code to think that all
previous "virtual files" are also found again. We can fake this by
returning bRC_Seen on each question to the check_file method of the
plugin which gets invoked for each file on the accurate Seen bitmap when
building the so called deleted files list.