bsock: allow sending long messages
While API modes 0 and 1 normally send data as small chunks (often line
based), API mode 2 (JSON) collects the result of a bcommand, forms a
JSON object and sends it. This can result in large messages.

The maximum valid size of the result of a bcommand used to be 4 MB.
The size of "llist jobs limit=1" is already close to 1 KB.
So before this patch, "llist jobs" silently fails if the Bareos
Director returns more than 5000 jobs.

With this patch, the maximum length of a bcommand result is int32 (2GB).