Where = / is passed as empty string.
We try to determine the name of the restore object based on the name of
the VM disk we are restoring. When the where is set to something other
then '/' the restorepkt.ofname is relative to the given where path but
when no where is specified its not and we should not try to determine
the releative path but just use the restorepkt.ofname directly.