Fix Director crash when migrating an empty job
When there is nothing to do for a Migration or Copy the migration_init()
function sets the Job exit code already to JS_Terminated. This patch
checks after the migration_init() if the Job Status is alread
JS_Terminated and then call migration_cleanup() directly.

I also added an explicit check in do_actual_migration() that we have a
mig_jcr setup as the rest of the code expect this to be set and for any
Migration or Copy Job that needs to do some actual work this must also
be set so the extra check is mostly to make sure it is.

The printing of the New JobId is also fixed to show *None* instead of
<NULL>.

Fixes #544: Director crash (SIGSEGV 11) when migrating an empty job