Fix random crashes on sd
The block variable was set to the dcr->block, but that can be altered in
the call to dev->set_label_blocksize(dcr).

When that happens, the code goes on with the wrong block.
We removed the whole local variable as it makes no sense and is only
referenced 3 times when calling empty_block()

Fixes #414: Bareos storage daemon crashes during backups
Fixse #483: bareos-sd crash during backup
Fixes #522: storage daemon crashes ocassionally when starting a new job
Fixes #552: SD crashes in -current

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>