fix(config): Add an hidden conf option to keep / in root paths
The request path computation functions were previously translating a
request to the '/' resource into an empty path, thus creating a
behavior that was unsuited to the CDMI backend, which requires the
ending / of a path accessing a directory to be present for metadata
reading requests.

This issue is worked around by adding a new cdmi-specific hidden option that
influences the behavior of this path computing function, preserving the / for
requests on the root path. This option is hidden because it should not be
visible by an user, as it's a backend-specific internal behavior configuration.

In order to automatically set this hidden option, the function dpl_backend_find
was renamed into dpl_backend_set and now sets both the backend and the hidden
options related to it.

Fixes #208