migration control jobs don't count for concurrency
Migration control jobs do not touch the storage in any way
so they do not need to be counted when checking the maximum
concurrent jobs for storages.

Also did a cleanup of the the code and comments along the way.

Fixes #528: Migration job hangs waiting for waiting for max Storage jobs

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>