We shouldn't do spooling on copy/migrate jobs.
We clone the Migrate/Copy Job to be an clone of the original Job and as
such we also clone things like Spool Data and Spool Size. This patch
resets the cloned Job to use the Spool Data and Spool Size setting of
the actual Migrate/Copy control Job.

Fixes #555: Copy jobs don't respect "Spool Data = no"