Finally fix LDFLAGS position.
The Makefiles for most binaries had put $(LDFLAGS) /before/ all of the local includes, giving /usr/local/lib more priority over ../libs, etc. This causes the compiler to try to link the programs with libs from the currently installed version, not the version being built.

This is a problem inherited from bacula - see the following bug reports in FreeBSD's ports:
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=192526
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=193641