migration: Only update Job type on success.
The attempt to migrate a job whose job resource is not defined in the
current director configuration (because the resource has been removed)
fails with an error like

	Fatal error: Previous Job resource not found for "rur_system_job"

However, in this case the "Type" field of the previous job is still
updated from "B" (backup) to "M" (Migration). This is wrong because
nothing was actually migrated.

This simple patch changes migration_cleanup() to mark the previous
job as migrated only if the migration job terminated successfully.

Signed-off-by: Andre Noll <maan@tuebingen.mpg.de>